move some tests that depend on tsserver into tsserver/
