+++ src/testRunner/tests.ts
@@ -53,8 +53,6 @@ import "./unittests/services/extract/constants";
 import "./unittests/services/extract/functions";
 import "./unittests/services/extract/symbolWalker";
 import "./unittests/services/extract/ranges";
-import "./unittests/services/findAllReferences";
-import "./unittests/services/goToDefinition";
 import "./unittests/services/hostNewLineSupport";
 import "./unittests/services/languageService";
 import "./unittests/services/organizeImports";
@@ -150,12 +148,14 @@ import "./unittests/tsserver/events/projectLoading";
 import "./unittests/tsserver/events/projectUpdatedInBackground";
 import "./unittests/tsserver/exportMapCache";
 import "./unittests/tsserver/externalProjects";
+import "./unittests/tsserver/findAllReferences";
 import "./unittests/tsserver/forceConsistentCasingInFileNames";
 import "./unittests/tsserver/formatSettings";
 import "./unittests/tsserver/getApplicableRefactors";
 import "./unittests/tsserver/getEditsForFileRename";
 import "./unittests/tsserver/getExportReferences";
 import "./unittests/tsserver/getFileReferences";
+import "./unittests/tsserver/goToDefinition";
 import "./unittests/tsserver/importHelpers";
 import "./unittests/tsserver/inlayHints";
 import "./unittests/tsserver/inferredProjects";
